+++ packages/article-main-comment/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.67.47](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.67.46...@times-components/article-main-comment@2.67.47) (2023-07-25)
+
+**Note:** Version bump only for package @times-components/article-main-comment
+
+
+
+
+
 ## [2.67.46](https://github.com/newsuk/times-components/compare/@times-components/article-main-comment@2.67.45...@times-components/article-main-comment@2.67.46) (2023-07-25)
 
 **Note:** Version bump only for package @times-components/article-main-comment